+++ ts/packages/cli/src/commands/interactive.ts
@@ -2,12 +2,10 @@
 // Licensed under the MIT License.
 
 import { Args, Command, Flags } from "@oclif/core";
-import readline from "readline/promises";
 import {
     getBuiltinSchemaNames,
     getCacheFactory,
     processCommand,
-    processRequests,
     getPrompt,
     initializeCommandHandlerContext,
     CommandHandlerContext,
@@ -15,7 +13,10 @@ import {
 } from "agent-dispatcher/internal";
 import inspector from "node:inspector";
 import { getChatModelNames } from "aiclient";
-import { createConsoleClientIO } from "agent-dispatcher/helpers/console";
+import {
+    processRequests,
+    createConsoleClientIO,
+} from "agent-dispatcher/helpers/console";
 
 const modelNames = await getChatModelNames();
 
@@ -63,10 +64,6 @@ export default class Interactive extends Command {
         if (flags.debug) {
             inspector.open(undefined, undefined, true);
         }
-        const stdio = readline.createInterface({
-            input: process.stdin,
-            output: process.stdout,
-        });
 
         let context: CommandHandlerContext | undefined;
 
@@ -80,7 +77,7 @@ export default class Interactive extends Command {
                 explainer: { name: flags.explainer },
                 persistSession: !flags.memory,
                 enableServiceHost: true,
-                clientIO: createConsoleClientIO(stdio),
+                clientIO: createConsoleClientIO(),
             });
 
             if (args.input) {
@@ -92,7 +89,6 @@ export default class Interactive extends Command {
 
             await processRequests<CommandHandlerContext>(
                 getPrompt,
-                stdio,
                 processCommand,
                 context,
             );
@@ -100,7 +96,6 @@ export default class Interactive extends Command {
             if (context) {
                 await closeCommandHandlerContext(context);
             }
-            stdio?.close();
         }
 
         // Some background network (like monogo) might keep the process live, exit explicitly.
